avoid circular dependency check
