+++ packages/core/src/tracing/index.ts
@@ -5,7 +5,7 @@ export { SentrySpan } from './sentrySpan';
 export { SentryNonRecordingSpan } from './sentryNonRecordingSpan';
 export { Transaction } from './transaction';
 // eslint-disable-next-line deprecation/deprecation
-export { getActiveTransaction, getActiveSpan, getSpanDescendants } from './utils';
+export { getActiveTransaction, getActiveSpan } from './utils';
 export {
   setHttpStatus,
   getSpanStatusFromHttpCode,
